python/requirements.txt
python-dotenv
paho-mqtt
streamlit
pandas
